+++ legacy/builder/phases/sizer.go
@@ -22,7 +22,6 @@ import (
 	"strconv"
 
 	"github.com/arduino/arduino-cli/arduino/builder/utils"
-	"github.com/arduino/arduino-cli/legacy/builder/builder_utils"
 	"github.com/arduino/arduino-cli/legacy/builder/types"
 	"github.com/arduino/go-properties-orderedmap"
 	"github.com/pkg/errors"
@@ -50,7 +49,7 @@ func (s *Sizer) Run(ctx *types.Context) error {
 }
 
 func checkSizeAdvanced(ctx *types.Context, properties *properties.Map) error {
-	command, err := builder_utils.PrepareCommandForRecipe(properties, "recipe.advanced_size.pattern", false)
+	command, err := utils.PrepareCommandForRecipe(properties, "recipe.advanced_size.pattern", false)
 	if err != nil {
 		return errors.New(tr("Error while determining sketch size: %s", err))
 	}
@@ -182,7 +181,7 @@ func checkSize(ctx *types.Context, buildProperties *properties.Map) error {
 }
 
 func execSizeRecipe(ctx *types.Context, properties *properties.Map) (textSize int, dataSize int, eepromSize int, resErr error) {
-	command, err := builder_utils.PrepareCommandForRecipe(properties, "recipe.size.pattern", false)
+	command, err := utils.PrepareCommandForRecipe(properties, "recipe.size.pattern", false)
 	if err != nil {
 		resErr = fmt.Errorf(tr("Error while determining sketch size: %s"), err)
 		return
